Image for products Insert data db helper

+++ app/src/main/java/com/csis3175/walmarket/entity/OrderItem.java
@@ -4,7 +4,7 @@ public class OrderItem {
     private Integer orderId;
     private Integer itemId;
     private Double price;
-    private Double quantity;
+    private Integer quantity;
 
     public Integer getOrderId() {
         return orderId;
@@ -30,11 +30,11 @@ public class OrderItem {
         this.price = price;
     }
 
-    public Double getQuantity() {
+    public Integer getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Double quantity) {
+    public void setQuantity(Integer quantity) {
         this.quantity = quantity;
     }
 }
